+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/util/ThrowingFunction.java
@@ -1,0 +1,8 @@
+package uk.ac.bbsrc.tgac.miso.core.util;
+
+@FunctionalInterface
+public interface ThrowingFunction<T, R, E extends Exception> {
+
+  public R apply(T t) throws E;
+
+}